+++ Charts/Classes/Renderers/ChartYAxisRenderer.swift
@@ -182,7 +182,7 @@ public class ChartYAxisRenderer: ChartAxisRendererBase
         CGContextSetLineWidth(context, _yAxis.axisLineWidth);
         if (_yAxis.axisLineDashLengths != nil)
         {
-            CGContextSetLineDash(context, _yAxis.axisLineDashPhase, _yAxis.axisLineDashLengths, _yAxis.axisLineDashLengths.count);
+            CGContextSetLineDash(context, _yAxis.axisLineDashPhase, _yAxis.axisLineDashLengths, UInt(_yAxis.axisLineDashLengths.count));
         }
         else
         {
@@ -254,7 +254,7 @@ public class ChartYAxisRenderer: ChartAxisRendererBase
         CGContextSetLineWidth(context, _yAxis.gridLineWidth);
         if (_yAxis.gridLineDashLengths != nil)
         {
-            CGContextSetLineDash(context, _yAxis.gridLineDashPhase, _yAxis.gridLineDashLengths, _yAxis.gridLineDashLengths.count);
+            CGContextSetLineDash(context, _yAxis.gridLineDashPhase, _yAxis.gridLineDashLengths, UInt(_yAxis.gridLineDashLengths.count));
         }
         else
         {
@@ -317,7 +317,7 @@ public class ChartYAxisRenderer: ChartAxisRendererBase
             CGContextSetLineWidth(context, l.lineWidth);
             if (l.lineDashLengths != nil)
             {
-                CGContextSetLineDash(context, l.lineDashPhase, l.lineDashLengths!, l.lineDashLengths!.count);
+                CGContextSetLineDash(context, l.lineDashPhase, l.lineDashLengths!, UInt(l.lineDashLengths!.count));
             }
             else
             {
